refactor: move formats into web/formats
